fix: cast loop iterable variables to int32 in CUDA device functions
Cast loop iterable variables (n, stage_count, accumulator_length, n_drivers, etc.) to int32 in factory closures and device code to prevent downstream computations from being upcast to 64-bit.

Modified files:
- src/cubie/integrators/algorithms/*.py
- src/cubie/outputhandling/*.py
- src/cubie/outputhandling/summarymetrics/peaks.py
- src/cubie/outputhandling/summarymetrics/negative_peaks.py
- tests/all_in_one.py

+++ src/cubie/outputhandling/save_state.py
@@ -7,7 +7,7 @@ during integration.
 
 from typing import Callable, Sequence, Union
 
-from numba import cuda
+from numba import cuda, int32
 from numpy.typing import ArrayLike
 
 from cubie.cuda_simsafe import compile_kwargs
@@ -58,8 +58,8 @@ def save_state_factory(
     ``output_observables_slice``, and ``output_counters_slice`` in place.
     """
     # Extract sizes from heights object
-    nobs = len(saved_observable_indices)
-    nstates = len(saved_state_indices)
+    nobs = int32(len(saved_observable_indices))
+    nstates = int32(len(saved_state_indices))
 
     @cuda.jit(
         device=True,
@@ -124,7 +124,7 @@ def save_state_factory(
             output_states_slice[nstates] = current_step
         
         if save_counters:
-            for i in range(4):
+            for i in range(int32(4)):
                 output_counters_slice[i] = current_counters[i]
         # no cover: stop
